pyproject.toml
[tool.poetry]
name = "msys2-devtools"
version = "0.1.0"
description = ""
authors = ["Christoph Reiter <reiter.christoph@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
netaddr = "^0.8.0"
tabulate = "^0.9.0"
requests = "^2.28.2"
zstandard = "^0.21.0"
pyyaml = "^6.0"
pydantic = "^2.0"

[tool.poetry.scripts]
msys2-srcinfo-cache = 'msys2_devtools.srcinfo_cache:run'
msys2-pypi-cache = 'msys2_devtools.pypi_cache:run'

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
flake8 = "^6.0.0"
types-PyYAML = "^6.0.12.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
